requirements.txt
fastapi
uvicorn
pydantic 
python-dotenv
langchain
langchain-community
langchain-huggingface
faiss-cpu
sentence-transformers
